refactored order model according to Tori advice

+++ server/db/models/Product.js
@@ -1,14 +1,14 @@
-const Sequelize = require('sequelize');
-const db = require('../db');
+const Sequelize = require("sequelize");
+const db = require("../db");
 
-const Product = db.define('product', {
+const Product = db.define("product", {
   name: {
     type: Sequelize.STRING,
     unique: true,
     allowNull: false,
   },
   price: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.INTEGER,
     allowNull: false,
   },
   quantity: {
@@ -18,11 +18,11 @@ const Product = db.define('product', {
   imageURL: {
     type: Sequelize.STRING,
     defaultValue:
-      'https://www.saashub.com/images/app/service_logos/16/1e550b6dd28e/large.png?1539661217',
+      "https://www.saashub.com/images/app/service_logos/16/1e550b6dd28e/large.png?1539661217",
   },
   category: {
-    type: Sequelize.ENUM('LOW_CALORIE', 'HIGH_CALORIE', 'LETHAL'),
-    defualtValue: 'HIGH_CALORIE',
+    type: Sequelize.ENUM("LOW_CALORIE", "HIGH_CALORIE", "LETHAL"),
+    defualtValue: "HIGH_CALORIE",
   },
 });
